backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pytesseract
PyMuPDF
scikit-learn
pandas
pydantic
pydantic-settings
pydantic[email]
Pillow
python-multipart
python-docx
httpx
opencv-python-headless
numpy